+++ src/app/components/statuses/add-statuses-group-dialog.component.spec.ts
@@ -1,11 +1,11 @@
 import { TaskStatus } from '@aneoconsultingfr/armonik.api.angular';
 import { TestBed } from '@angular/core/testing';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { EditStatusesGroupDialogComponent } from './edit-status-group-dialog.component';
-import { TasksStatusesGroup } from '../types';
+import { AddStatusesGroupDialogComponent } from './add-statuses-group-dialog.component';
+import { TasksStatusesGroup } from '../../dashboard/types';
 
 describe('', () => {
-  let component: EditStatusesGroupDialogComponent;
+  let component: AddStatusesGroupDialogComponent;
 
   const mockMatDialogRef = {
     close: jest.fn()
@@ -14,14 +14,13 @@ describe('', () => {
   beforeEach(() => {
     component = TestBed.configureTestingModule({
       providers: [
-        EditStatusesGroupDialogComponent,
+        AddStatusesGroupDialogComponent,
         { provide: MatDialogRef, useValue: mockMatDialogRef },
         { provide: MAT_DIALOG_DATA, useValue: {
-          statuses: [{ name: 'result', value: 'the-result' }],
-          group: { name: 'status', color: 'green', statuses: [TaskStatus.TASK_STATUS_CANCELLED, TaskStatus.TASK_STATUS_COMPLETED]}
+          statuses: [{ name: 'result', value: 'the-result' }]
         } }
       ]
-    }).inject(EditStatusesGroupDialogComponent);
+    }).inject(AddStatusesGroupDialogComponent);
   });
 
   it('should create', () => {
@@ -31,7 +30,6 @@ describe('', () => {
   it('should init', () => {
     component.ngOnInit();
     expect(component.statuses).toEqual([{ name: 'result', value: 'the-result' }]);
-    expect(component.group).toEqual({ name: 'status', color: 'green', statuses: [TaskStatus.TASK_STATUS_CANCELLED, TaskStatus.TASK_STATUS_COMPLETED]});
   });
 
   it('should close with result on submit', () => {
