+++ src/app/components/statuses/edit-status-group-dialog.component.spec.ts
@@ -1,11 +1,11 @@
 import { TaskStatus } from '@aneoconsultingfr/armonik.api.angular';
 import { TestBed } from '@angular/core/testing';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { AddStatusesGroupDialogComponent } from './add-statuses-group-dialog.component';
-import { TasksStatusesGroup } from '../types';
+import { EditStatusesGroupDialogComponent } from './edit-status-group-dialog.component';
+import { TasksStatusesGroup } from '../../dashboard/types';
 
 describe('', () => {
-  let component: AddStatusesGroupDialogComponent;
+  let component: EditStatusesGroupDialogComponent;
 
   const mockMatDialogRef = {
     close: jest.fn()
@@ -14,13 +14,14 @@ describe('', () => {
   beforeEach(() => {
     component = TestBed.configureTestingModule({
       providers: [
-        AddStatusesGroupDialogComponent,
+        EditStatusesGroupDialogComponent,
         { provide: MatDialogRef, useValue: mockMatDialogRef },
         { provide: MAT_DIALOG_DATA, useValue: {
-          statuses: [{ name: 'result', value: 'the-result' }]
+          statuses: [{ name: 'result', value: 'the-result' }],
+          group: { name: 'status', color: 'green', statuses: [TaskStatus.TASK_STATUS_CANCELLED, TaskStatus.TASK_STATUS_COMPLETED]}
         } }
       ]
-    }).inject(AddStatusesGroupDialogComponent);
+    }).inject(EditStatusesGroupDialogComponent);
   });
 
   it('should create', () => {
@@ -30,6 +31,7 @@ describe('', () => {
   it('should init', () => {
     component.ngOnInit();
     expect(component.statuses).toEqual([{ name: 'result', value: 'the-result' }]);
+    expect(component.group).toEqual({ name: 'status', color: 'green', statuses: [TaskStatus.TASK_STATUS_CANCELLED, TaskStatus.TASK_STATUS_COMPLETED]});
   });
 
   it('should close with result on submit', () => {
