+++ src/app/components/statuses/manage-groups-dialog.component.html
@@ -1,4 +1,4 @@
-<h2 mat-dialog-title i18n="Dialog title">Manage Groups</h2>
+<h2 mat-dialog-title i18n="Dialog title">Manage Statuses Groups</h2>
 
 <mat-dialog-content>
   <ul class="groups" cdkDropListGroup>
@@ -8,13 +8,11 @@
           {{ group.name }}
         </h3>
         <div class="group-header-actions">
-          <button mat-icon-button (click)="openEditStatusGroupModal(group)">
+          <button mat-icon-button (click)="openEditStatusGroupModal(group)" i18n-matTooltip="Edit the group" matTooltip="Edit">
             <mat-icon [fontIcon]="getIcon('edit')"/>
-            <span class="sr-only" i18n="Edit the group">Edit {{ group.name }}</span>
           </button>
-          <button mat-icon-button (click)="onDelete(group)">
+          <button mat-icon-button (click)="onDelete(group)" i18n-matTooltip="Delete the group" matTooltip="Delete">
             <mat-icon [fontIcon]="getIcon('delete')"/>
-            <span class="sr-only" i18n="Delete the group">Delete {{ group.name }}</span>
           </button>
         </div>
       </div>
@@ -29,8 +27,9 @@
       </ul>
     </li>
     <div class="addGroup">
-      <button mat-fab (click)="openAddStatusGroupModal()" matTooltip="Add a group">
+      <button mat-stroked-button (click)="openAddStatusGroupModal()">
         <mat-icon aria-hidden="true" [fontIcon]="getIcon('add')"/>
+        <span i18n="Open a modal on click">Add a group</span>
       </button>
     </div>
   </ul>
